"Adds packages and translated into an architectural pattern MVVM"
