+++ app/src/main/java/com/mikerusetsky/appcinema/domain/Interactor.kt
@@ -1,0 +1,7 @@
+package com.mikerusetsky.appcinema.domain
+
+import com.mikerusetsky.appcinema.data.MainRepository
+
+class Interactor (val repo: MainRepository) {
+    fun getFilmsDB(): List<Film> = repo.filmsDataBase
+}
