+++ app/src/main/java/com/mikerusetsky/appcinema/view/fragments/CollectionsFragment.kt
@@ -1,4 +1,4 @@
-package com.mikerusetsky.appcinema
+package com.mikerusetsky.appcinema.view.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,7 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.mikerusetsky.appcinema.databinding.FragmentCollectionsBinding
-import com.mikerusetsky.appcinema.databinding.FragmentHomeBinding
+import com.mikerusetsky.appcinema.utils.AnimationHelper
 
 
 class CollectionsFragment : Fragment() {
