+++ app/src/main/java/com/mikerusetsky/appcinema/view/fragments/FavoritesFragment.kt
@@ -1,4 +1,4 @@
-package com.mikerusetsky.appcinema
+package com.mikerusetsky.appcinema.view.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,7 +7,11 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.mikerusetsky.appcinema.databinding.FragmentFavoritesBinding
-
+import com.mikerusetsky.appcinema.domain.Film
+import com.mikerusetsky.appcinema.utils.AnimationHelper
+import com.mikerusetsky.appcinema.view.rv_adapters.FilmListRecyclerAdapter
+import com.mikerusetsky.appcinema.view.MainActivity
+import com.mikerusetsky.appcinema.view.rv_adapters.TopSpacingItemDecoration
 
 
 class FavoritesFragment : Fragment() {
