+++ app/src/main/java/com/mikerusetsky/appcinema/view/fragments/WatchLaterFragment.kt
@@ -1,12 +1,12 @@
-package com.mikerusetsky.appcinema
+package com.mikerusetsky.appcinema.view.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.mikerusetsky.appcinema.databinding.FragmentHomeBinding
 import com.mikerusetsky.appcinema.databinding.FragmentWatchlaterBinding
+import com.mikerusetsky.appcinema.utils.AnimationHelper
 
 class WatchLaterFragment : Fragment() {
     private lateinit var binding3: FragmentWatchlaterBinding
