+++ app/src/main/java/com/mikerusetsky/appcinema/viewmodel/HomeFragmentViewModel.kt
@@ -1,0 +1,17 @@
+package com.mikerusetsky.appcinema.viewmodel
+
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import com.mikerusetsky.appcinema.App
+import com.mikerusetsky.appcinema.domain.Film
+import com.mikerusetsky.appcinema.domain.Interactor
+
+class HomeFragmentViewModel : ViewModel() {
+    val filmsListLiveData = MutableLiveData<List<Film>>()
+    //Инициализируем интерактор
+    private var interactor: Interactor = App.instance.interactor
+    init {
+        val films = interactor.getFilmsDB()
+        filmsListLiveData.postValue(films)
+    }
+}
