+++ app/src/main/res/layout/activity_splash.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/background_dark"
-    tools:context=".SplashActivity">
+    tools:context=".view.SplashActivity">
 
     <com.airbnb.lottie.LottieAnimationView
         android:id="@+id/splash_view"
