+++ app/src/main/res/layout/fragment_favorites.xml
@@ -8,7 +8,7 @@
     tools:visibility="visible"
     android:id="@+id/favorites_fragment_root"
     android:background="@drawable/favourites_screen"
-    tools:context=".FavoritesFragment">
+    tools:context=".view.fragments.FavoritesFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/favorites_recycler"
