+++ app/src/main/res/layout/fragment_watchlater.xml
@@ -7,6 +7,6 @@
     android:background="@drawable/watchlater_sccreen"
     android:visibility="invisible"
     tools:visibility="visible"
-    tools:context=".WatchLaterFragment">
+    tools:context=".view.fragments.WatchLaterFragment">
 
 </FrameLayout>
